Start using NixOS and home-manager as my daily.
This *very large* commit is a squash of hundreds of small commits I've
made over the last month with evaluating the usability of NixOS [1] and
home-manager [2]. I have decided to make the plunge of go full time with
this setup.

I made the choice to explore NixOS since in my dotfiles for Debian I had
a master `update-system.sh` script that was trying to make my
configuration immutable. I had functions to install applications from
sources with pinned version numbers and lists of packages to be
installed via apt.

NixOS takes this idea and build an entire operating system around it. At
this point I am a fan. My main concern is losing the ability to share
dotfiles with colleagues if I move configuration into home-manager's
declarative model. I may end up not using this feature of home-manager,
only time will tell.

My migration is in now way complete, and there will be many change in
the next few weeks as I continue to workout my configuration.

[1] https://nixos.org/
[2] https://github.com/rycee/home-manager