feat(gpg): cut extra gpg-agent configuration
This commit removes the `oh-my-zsh` `gpg-agent` plugin and an unused
configuration file.

I have a hunch that this plugin has something to due with all the issue
I am having with gpg-agent crashing on me. I configure nixos to start
and manage my gpg-agent so this plugin _should_ be redundant at best,
and conflicting at worse.