feat(env): add work-in-progress scripts to path
This change allows me to more easily iterate on helpful scripts without
the pressure to ensure the scripts are of good enough quality to add to
this repository.

With that said, I do try and make every effort to have scripts live in
this directory as little as possible. Its mostly for initial development
of scripts.