feat(docker): add pull through registry mirror
This change allows docker to pull from a local cache instead of pulling
from upstreams all the time. This is nice because if the mirror is not
available, docker will just fall through to the normal upstream.

There are plans to host this within my home-lab once it is up and
running.