feat: add ZSH binding and func to pick git commit
The commit allows me to _very_ easily pick a git commit from a nicely
formatted list and drop it right into my current command.

The impetus for this command was my quest to short the time and lessen
the effort needed to apply a set of changes in git to a previous commit.