feat(zsh): add function to edit file or change dir
This little function will allow me better ergonomics and the ability to
use a single letter to either edit a file, or move into a directory.

This same end goal can be accomplished by using zsh's autocd options,
along with settings alias for extensions. I choose not to do that to
keep everything simple.