feat(nvim): source a spellfix defined in code
This will make it easier to share a common spellfile between all
machines I manage. Settings multiple spellfiles separated by commas is a
supported usage, which is nice! Using `zg` will add the word to the
first writeable spell file, which will be the standard spellfile on that
computer. I'll just need to periodically check that and add words to the
file defined in this repo.