feat: add TF code for EC2 aarch64 builder
This is useful when building large amount of aarch64-linux packages.
Mostly for kernel rebuilds. To build locally I either ship it to a
RaspberryPi, which will take forever, or build it on my build server
with `binftm`, which is quicker, but not fast.

Using a large EC2 instance builds a kernel in a few minutes,