fix(vim-tmux-navigator): use recommend bindings
This update allows my movement bindings to work with `vimdiff` mode.

I noticed that I was unable to move between splits when using vimdiff. I
went digging and found a new set of keybinding were recommend in the
README [1].

I choose not to use the plugin method since I am using a non-default
keybinding and this felt more clear.

I did attempt to use the plugin, even a more recent version, but could
not get my new keybindings to override the keybindings defined in code.

[1]: https://github.com/christoomey/vim-tmux-navigator/blob/6a1e58c3ca3bc7acca36c90521b3dfae83b2a602/README.md#add-a-snippet